code/requirements.txt
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2020.6.20
cffi==1.15.0
colorama==0.4.4
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
importlib-resources==5.4.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.0.1
prometheus-client==0.12.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.7.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==227
pywinpty==1.1.5
PyYAML==6.0
pyzmq==19.0.2
scikit-learn==1.0.1
scipy==1.7.1
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typing-extensions==3.10.0.2
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.0
widgetsnbextension==3.5.2
wincertstore==0.2
zipp==3.6.0
